pyproject.toml
[project]
name = "api-test-django"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "takumi-13", email = "takumi.gameover13@gmail.com" }
]
dependencies = [
    "django>=5.1.4",
    "django-environ>=0.11.2",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.6.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api_test_django"]

# Lint/Formatter非依存のルール
[tool.ruff]
# 除外したいファイル
# excludeを設定するとデフォルトで設定されている.venv配下などの設定が上書きされるため要注意
extend-exclude = ["src/something.py"]
# 行文字数
line-length = 88
# インデント幅
indent-width = 4

# Linterのルール
[tool.ruff.lint]
# Lintルール
# https://docs.astral.sh/ruff/rules/　に記載されているルールのうち、Codeが前方一致するルールが適用されます
select = ["E", "F", "I", "PLR"]
# 除外するLintルール
# selectで選択したルールから除外したいものを指定できます
ignore = ["PLR2004"]
# auto fixの対象にするルール
fixable = ["ALL"]
# auto fixから除外するルール
# editor.formatOnSaveを使う際に自動でFixされると困るものを指定できます
# unfixable = ["F401"]

# Formatterのルール
[tool.ruff.format]
# 文字列をダブルクォートに統一
quote-style = "double"
# インデントをタブ文字(\t)からスペースに変換
indent-style = "space"

select = [
    "I",
]
